Move plugin directory from model to project root.
